Editorial: simplify early errors
Prior to this commit, the early errors for
ClassStaticInitializationBlock were not applied when nesting crossed
"function or `static` initialization block boundaries." In the latter
case (that is: nesting within a "static" initialization block), the rule
applies recursively, triggering the early error despite the exception.

Remove the exception to simplify the rule's definition.